MFC r295823
Modified the use of bxe_grc_dump() function so that it can be invoked directly
at any potential error path, where a fwdump is needed.
The fwdump (a.k.a grcdump) is stored in a driver buffer.
The sysctl grcdump_done indicates if a fwdump was taken and waiting to be retrieved.
The sysctl trigger_grcdump can be used to manually trigger a fwdump.

Approved by:re (marius)


git-svn-id: https://svn.freebsd.org/base/stable/10@296061 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f