MFC 296932: pf: Improve forwarding detection
When we guess the nature of the outbound packet (output vs. forwarding) we need
to take bridges into account. When bridging the input interface does not match
the output interface, but we're not forwarding. Similarly, it's possible for the
interface to actually be the bridge interface itself (and not a member interface).


git-svn-id: https://svn.freebsd.org/base/stable/10@297429 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f