Don't wakeup the fdc worker thread once a second when idle.
The fdc worker thread was using a one second timeout while waiting for
a new bio to arrive or for the device to detach.  However, the driver
already does a wakeup when queueing a new bio or asking the thread to
detach, so the timeout only served to waste CPU time waking up the
thread once a second just so it could go right back to sleep.  Use an
infinite timeout instead.

Discussed with:	phk
Sponsored by:	Netflix


git-svn-id: https://svn.freebsd.org/base/head@297568 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f